Added support for using menus other that <menu>
Identified that some use cases in Drupal are better off not using contextual menus because of modal windows.
It is sometimes therefore better to use a <ul> or similar construct in order to build your menu structure.
In order to allow for a similar look & feel between contextual menus and non-contextual menus, this patch
allows for any tag to be used as a menu structure (by tagging it as .gearmenu-element--list-menu or by
specifying it in the selector) and for it to be associated with an element by using the data-list-contextmenu
custom HTML attribute (used in the same way as the contextmenu attribute).